Push GradlDistributionInfo implementation details to internal subclass
